+++ synthmri/models/__init__.py
@@ -1,4 +1,4 @@
 from .unet import build_unet, count_parameters
-from .vae import DummyVAE, VAEWrapper, load_vae
+from .vae import DummyVAE, VAEWrapper, decoder_state_dict, load_decoder_weights, load_vae
 
-__all__ = ["DummyVAE", "VAEWrapper", "build_unet", "count_parameters", "load_vae"]
+__all__ = ["DummyVAE", "VAEWrapper", "build_unet", "count_parameters", "decoder_state_dict", "load_decoder_weights", "load_vae"]